+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas/Modules/AzMap.cs
@@ -10,12 +10,14 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
     /// </summary>
     public interface IAtlasMap
     {
+        #region Camera
+
         /// <summary>
         /// Returns the map camera's current properties.
         /// </summary>
         /// <param name="map"></param>
         /// <returns></returns>
-        Task<MapCamera> GetCamera(Map map);
+        ValueTask<MapCamera> GetCamera(Map map);
 
         /// <summary>
         /// Set the camera of the map control with an animated transition. 
@@ -27,13 +29,97 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
         /// <param name="animation"></param>
         /// <returns></returns>
         ValueTask SetCamera(Map map, CameraOptions? camera, CameraBoundsOptions? cameraBounds = null, AnimationOptions? animation = null);
+
+        #endregion
+
+        #region Service
+
+        /// <summary>
+        /// Get the map service options.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <returns></returns>
+        ValueTask<ServiceOptions> GetServiceOptions(Map map);
+
+        /// <summary>
+        /// Set the map service options.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <param name="options"></param>
+        /// <param name="setAction"><see cref="SetOptionsAction"/>. Default is <see cref="SetOptionsAction.Update"/> </param>
+        /// <returns></returns>
+        ValueTask SetServiceOptions(Map map, ServiceOptions options, SetOptionsAction setAction = SetOptionsAction.Update);
+
+        #endregion
+
+        #region Style
+
+        /// <summary>
+        /// Get the map style options.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <returns></returns>
+        ValueTask<StyleOptions> GetStyle(Map map);
+
+        /// <summary>
+        /// Set the map style options.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <param name="options"></param>
+        /// <param name="setAction"><see cref="SetOptionsAction"/>. Default is <see cref="SetOptionsAction.Update"/> </param>
+        /// <returns></returns>
+        ValueTask SetStyle(Map map, StyleOptions options, SetOptionsAction setAction = SetOptionsAction.Update);
+
+        #endregion
+
+        #region Traffic
+
+        /// <summary>
+        /// Get the map traffic options.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <returns></returns>
+        ValueTask<TrafficOptions> GetTraffic(Map map);
+
+        /// <summary>
+        /// Set the traffic options for the map.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <param name="options"></param>
+        /// <param name="setAction"><see cref="SetOptionsAction"/>. Default is <see cref="SetOptionsAction.Update"/></param>
+        /// <returns></returns>
+        ValueTask SetTraffic(Map map, TrafficOptions options, SetOptionsAction setAction = SetOptionsAction.Update);
+
+        #endregion
+
+        #region UserInteraction
+
+        /// <summary>
+        /// Gets the map user interaction options.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <returns></returns>
+        ValueTask<UserInteractionOptions> GetUserInteraction(Map map);
+
+        /// <summary>
+        /// Sets the map user interaction options.
+        /// </summary>
+        /// <param name="map"></param>
+        /// <param name="options"></param>
+        /// <param name="setAction"><see cref="SetOptionsAction"/>. Default is <see cref="SetOptionsAction.Update"/></param>
+        /// <returns></returns>
+        ValueTask SetUserInteraction(Map map, UserInteractionOptions options, SetOptionsAction setAction = SetOptionsAction.Update);
+
+        #endregion
     }
 
     internal class AzMap(Lazy<Task<IJSObjectReference>> moduleTask) : IAtlasMap
     {
         private readonly Lazy<Task<IJSObjectReference>> moduleTask = moduleTask;
 
-        public async Task<MapCamera> GetCamera(Map map)
+        #region Camera
+
+        public async ValueTask<MapCamera> GetCamera(Map map)
         {
             return await moduleTask.InvokeAsyncTC<MapCamera>(GetJsInteropMethod(), map.MapReference);
         }
@@ -42,6 +128,64 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas.Modules
         {
             await moduleTask.InvokeVoidAsyncTC(GetJsInteropMethod(), map.MapReference, camera, cameraBounds, animation);
         }
+
+        #endregion
+
+        #region Service
+
+        public async ValueTask<ServiceOptions> GetServiceOptions(Map map)
+        {
+            return await moduleTask.InvokeAsyncTC<ServiceOptions>(GetJsInteropMethod(), map.MapReference);
+        }
+
+        public async ValueTask SetServiceOptions(Map map, ServiceOptions options, SetOptionsAction setAction = SetOptionsAction.Update)
+        {
+            await moduleTask.InvokeVoidAsyncTC(GetJsInteropMethod(), map.MapReference, options, setAction);
+        }
+
+        #endregion
+
+        #region Style
+
+        public async ValueTask<StyleOptions> GetStyle(Map map)
+        {
+            return await moduleTask.InvokeAsyncTC<StyleOptions>(GetJsInteropMethod(), map.MapReference);
+        }
+
+        public async ValueTask SetStyle(Map map, StyleOptions options, SetOptionsAction setAction = SetOptionsAction.Update)
+        {
+            await moduleTask.InvokeVoidAsyncTC(GetJsInteropMethod(), map.MapReference, options, setAction);
+        }
+
+        #endregion
+
+        #region Traffic
+
+        public async ValueTask<TrafficOptions> GetTraffic(Map map)
+        {
+            return await moduleTask.InvokeAsyncTC<TrafficOptions>(GetJsInteropMethod(), map.MapReference);
+        }
+
+        public async ValueTask SetTraffic(Map map, TrafficOptions options, SetOptionsAction setAction = SetOptionsAction.Update)
+        {
+            await moduleTask.InvokeVoidAsyncTC(GetJsInteropMethod(), map.MapReference, options, setAction);
+        }
+
+        #endregion
+
+        #region UserInteraction
+
+        public async ValueTask<UserInteractionOptions> GetUserInteraction(Map map)
+        {
+            return await moduleTask.InvokeAsyncTC<UserInteractionOptions>(GetJsInteropMethod(), map.MapReference);
+        }
+
+        public async ValueTask SetUserInteraction(Map map, UserInteractionOptions options, SetOptionsAction setAction = SetOptionsAction.Update)
+        {
+            await moduleTask.InvokeVoidAsyncTC(GetJsInteropMethod(), map.MapReference, options, setAction);
+        }
+
+        #endregion
 
         private static string GetJsInteropMethod([CallerMemberName] string name = "")
             => JsModule.Map.GetJsModuleMethod(name);