+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Atlas/scripts/atlas/common/index.ts
@@ -13,3 +13,8 @@ export interface MapObjectReference {
     type: string | undefined;
     jsReference: any;
 }
+
+export enum EditAction {
+    Update = 'update',
+    Replace = 'replace',
+}